chore: Refactor NAS
