--- cmd/suggestion/nasenvelopenet/Dockerfile
@@ -1,8 +1,0 @@
-FROM python:3.6
-
-ADD . /usr/src/app/github.com/kubeflow/katib
-WORKDIR /usr/src/app/github.com/kubeflow/katib/cmd/suggestion/nasenvelopenet
-RUN pip install --no-cache-dir -r requirements.txt
-ENV PYTHONPATH /usr/src/app/github.com/kubeflow/katib:/usr/src/app/github.com/kubeflow/katib/pkg/api/python
-
-ENTRYPOINT ["python", "-u", "main.py"]
