+++ examples/v1alpha1/NAS-Envelopenet-trainingcontainer/cell.py
