+++ examples/v1alpha1/NAS-Envelopenet-trainingcontainer/cell_init.py
