+++ examples/v1alpha1/NAS-Envelopenet-trainingcontainer/cifar10_input.py
