+++ examples/v1alpha1/NAS-Envelopenet-trainingcontainer/evaluate.py
