+++ examples/v1alpha1/NAS-Envelopenet-trainingcontainer/model_constructor.py
